refactor: fix types from ol update
